telegrambot/requirements.txt
python-telegram-bot
aiomqtt

